server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pycryptodome==3.19.0
pydantic==2.5.2
python-dotenv==1.0.0